Create temp file
Lots of redundant code here iwth the directory function, but create_temp_file
will now create a temporary file in a location of the user's suggesting.
If the user doesn't suggest a directory, the function defaults to the current directory and returns the path.

Obviously, this needs to be rearranged in some major ways. Very little code is shared.  These should be in two entirely different modules (Fileutils.Dir and Fileutils.File, perhaps?)

Low-Hanging Fruit: already_exists? is the same as does_directory_already_exist?